convert to int where required (follow-up 58fa49f8)
(cherry picked from commit fb58baba5ee1ddd76b8814b6633588247987042a)

+++ src/SAGisXPlanung/gui/widgets/QFeatureIdentify.py
@@ -100,7 +100,7 @@ class QFeatureIdentify(QXPlanInputElement, QtWidgets.QWidget, metaclass=XPlanung
             Korrekte Lösung wäre die Nutzung von SizePolicy's der zugrundeliegenden ComboBox, diese ist aber als
             privates Attribut `mComboBox` der Klasse `QgsFeaturePickerWidget` versteckt."""
         if event.type() == QEvent.Resize:
-            self.cbFeature.setMaximumWidth(self.mMapLayerComboBox.width() / 2)
+            self.cbFeature.setMaximumWidth(int(self.mMapLayerComboBox.width() / 2))
             return True
         return False
